LPS-156173 Move to layout-page-template-api to reuse
